Create GiftForm component to create more modular structure
